Add test_gemini_with_tools_parameters_set_to_is_annotated_with_none_as_default_value and mark it with skip until the issue is resolved
