+++ test/agentchat/test_dependancy_injection.py
@@ -262,29 +262,3 @@ class TestDependencyInjection:
         is_async: bool,
     ) -> None:
         await self._test_end2end(credentials_from_test_param, is_async)
-
-    @pytest.mark.skip(reason="This test is failing. We need to investigate the issue.")
-    @pytest.mark.gemini
-    def test_gemini_with_tools_parameters_set_to_is_annotated_with_none_as_default_value(
-        self,
-        credentials_gemini_pro: Credentials,
-    ) -> None:
-        agent = ConversableAgent(name="agent", llm_config=credentials_gemini_pro.llm_config)
-
-        user_proxy = UserProxyAgent(
-            name="user_proxy_1",
-            human_input_mode="NEVER",
-        )
-
-        mock = MagicMock()
-
-        @user_proxy.register_for_execution()
-        @agent.register_for_llm(description="Login function")
-        def login(
-            additional_notes: Annotated[Optional[str], "Additional notes"] = None,
-        ) -> str:
-            return "Login successful."
-
-        user_proxy.initiate_chat(agent, message="Please login", max_turns=2)
-
-        mock.assert_called_once()
